+++ lib/bloc_implementation.dart
@@ -1,4 +1,9 @@
 library bloc_implementation;
 
+/// File which exports all Classes the Package
+/// provides to the User.
+/// Usually this File is imported into a new Project
+
 export 'src/models/bloc.dart';
 export 'src/models/bloc_parent.dart';
+export 'src/exceptions/bloc_not_found_exception.dart';