+++ lib/src/models/bloc.dart
@@ -8,7 +8,7 @@ abstract class Bloc {
   Bloc();
 
   /// Stream Controller to use for the Stream
-  late final StreamController controller = StreamController();
+  final StreamController controller = StreamController();
 
   /// Method to dispose everything.
   /// Here you cancel all Listeners on